requirements.txt
fastapi
uvicorn
streamlit
requests
pandas
plotly
python-nmap
python-whois
